# packages/shared-types/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "InvitationCreate": {
        "additionalProperties": false,
        "properties": {
          "department": {
            "anyOf": [
              {
                "maxLength": 80,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Department"
          },
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "name": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "role_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "maxItems": 20,
            "minItems": 1,
            "title": "Role Ids",
            "type": "array"
          }
        },
        "required": [
          "name",
          "email",
          "role_ids"
        ],
        "title": "InvitationCreate",
        "type": "object"
      },
      "InvitationList": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/InvitationOut"
            },
            "title": "Data",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "data"
        ],
        "title": "InvitationList",
        "type": "object"
      },
      "InvitationOut": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "department": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Department"
          },
          "email": {
            "title": "Email",
            "type": "string"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "invited_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Invited By"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "roles": {
            "items": {
              "$ref": "#/components/schemas/RoleRef"
            },
            "title": "Roles",
            "type": "array"
          }
        },
        "required": [
          "id",
          "name",
          "email",
          "department",
          "roles",
          "invited_by",
          "expires_at",
          "created_at"
        ],
        "title": "InvitationOut",
        "type": "object"
      },
...
      "RoleCreate": {
        "additionalProperties": false,
        "properties": {
          "name": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "permissions": {
            "items": {
              "maxLength": 60,
              "pattern": "^[a-z_]+(\\.[a-z_]+){1,3}$",
              "type": "string"
            },
            "maxItems": 100,
            "minItems": 1,
            "title": "Permissions",
            "type": "array"
          }
        },
        "required": [
          "name",
          "permissions"
        ],
        "title": "RoleCreate",
        "type": "object"
      },
      "RoleList": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/RoleOut"
            },
            "title": "Data",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "data"
        ],
        "title": "RoleList",
        "type": "object"
      },
      "RoleOut": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_system": {
            "title": "Is System",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "permissions": {
            "items": {
              "type": "string"
            },
            "title": "Permissions",
            "type": "array"
          }
        },
        "required": [
          "id",
          "code",
          "name",
          "is_system",
          "permissions"
        ],
        "title": "RoleOut",
        "type": "object"
      },
      "RolePatch": {
        "additionalProperties": false,
        "properties": {
          "name": {
            "anyOf": [
              {
                "maxLength": 120,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "permissions": {
            "anyOf": [
              {
                "items": {
                  "maxLength": 60,
                  "pattern": "^[a-z_]+(\\.[a-z_]+){1,3}$",
                  "type": "string"
                },
                "maxItems": 100,
                "minItems": 1,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Permissions"
          }
        },
        "title": "RolePatch",
        "type": "object"
      },
      "RoleRef": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "code",
          "name"
        ],
        "title": "RoleRef",
        "type": "object"
      },
...
      "StaffMember": {
        "properties": {
          "department": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Department"
          },
          "email": {
            "title": "Email",
            "type": "string"
          },
          "has_pin": {
            "title": "Has Pin",
            "type": "boolean"
          },
          "joined_at": {
            "format": "date-time",
            "title": "Joined At",
            "type": "string"
          },
          "last_sign_in_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Sign In At"
          },
          "mfa_enabled": {
            "title": "Mfa Enabled",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "roles": {
            "items": {
              "$ref": "#/components/schemas/RoleRef"
            },
            "title": "Roles",
            "type": "array"
          },
          "status": {
            "enum": [
              "active",
              "deactivated"
            ],
            "title": "Status",
            "type": "string"
          },
          "user_id": {
            "format": "uuid",
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "user_id",
          "name",
          "email",
          "department",
          "status",
          "roles",
          "has_pin",
          "mfa_enabled",
          "last_sign_in_at",
          "joined_at"
        ],
        "title": "StaffMember",
        "type": "object"
      },
      "StaffPatch": {
        "additionalProperties": false,
        "properties": {
          "department": {
            "anyOf": [
              {
                "maxLength": 80,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Department"
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 120,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          }
        },
        "title": "StaffPatch",
        "type": "object"
      },
      "StaffRoles": {
        "additionalProperties": false,
        "properties": {
          "role_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "maxItems": 20,
            "minItems": 1,
            "title": "Role Ids",
            "type": "array"
          }
        },
        "required": [
          "role_ids"
        ],
        "title": "StaffRoles",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Diyneco API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/roles": {
      "get": {
        "operationId": "list_roles_api_v1_roles_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleList"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Roles",
        "tags": [
          "hotel"
        ]
      },
      "post": {
        "operationId": "create_role_api_v1_roles_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoleCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Role",
        "tags": [
          "staff"
        ]
      }
    },
    "/api/v1/roles/{role_id}": {
      "patch": {
        "operationId": "patch_role_api_v1_roles__role_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "role_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Role Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RolePatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Patch Role",
        "tags": [
          "staff"
        ]
      }
    },
...
    "/api/v1/staff/invitations": {
      "get": {
        "operationId": "list_invitations_api_v1_staff_invitations_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvitationList"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Invitations",
        "tags": [
          "staff"
        ]
      },
      "post": {
        "operationId": "invite_staff_api_v1_staff_invitations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvitationCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvitationOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Invite Staff",
        "tags": [
          "staff"
        ]
      }
    },
    "/api/v1/staff/invitations/{invitation_id}": {
      "delete": {
        "operationId": "cancel_invitation_api_v1_staff_invitations__invitation_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "invitation_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Invitation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Invitation",
        "tags": [
          "staff"
        ]
      }
    },
    "/api/v1/staff/{user_id}": {
      "patch": {
        "operationId": "patch_staff_api_v1_staff__user_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "User Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StaffPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StaffMember"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Patch Staff",
        "tags": [
          "staff"
        ]
      }
    },
    "/api/v1/staff/{user_id}/deactivate": {
      "post": {
        "operationId": "deactivate_staff_api_v1_staff__user_id__deactivate_post",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "User Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StaffMember"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Deactivate Staff",
        "tags": [
          "staff"
        ]
      }
    },
    "/api/v1/staff/{user_id}/pin/reset": {
      "post": {
        "operationId": "reset_staff_pin_api_v1_staff__user_id__pin_reset_post",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "User Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reset Staff Pin",
        "tags": [
          "staff"
        ]
      }
    },
    "/api/v1/staff/{user_id}/roles": {
      "put": {
        "operationId": "set_staff_roles_api_v1_staff__user_id__roles_put",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "User Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StaffRoles"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StaffMember"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Staff Roles",
        "tags": [
          "staff"
        ]
      }
    }
  }
}